pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Django boilerplate"
authors = ["Vinicius Pedrini <viniciusvmpz@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
django-simple-history = "^3.3.0"
django-allauth = "^0.54.0"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.3"}
pre-commit = "^3.3.3"
whitenoise = "^6.5.0"
django-compressor = "^4.4"
django-htmx = "^1.16.0"
pillow = "^10.0.0"
openai = "^0.27.9"
tiktoken = "^0.4.0"
pip = "^23.2.1"
ipykernel = "^6.25.1"
pgvector = "^0.2.1"
psycopg2-binary = "^2.9.7"
dj-database-url = "^2.1.0"


[tool.poetry.group.dev.dependencies]
django-browser-reload = "^1.11.0"
djlint = "^1.32.1"
ruff = "^0.0.285"
black = "^23.7.0"
pytest-django = "^4.5.2"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "app.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"
python_files = ["tests.py", "test_*.py",  "*_tests.py"]
